// packages/features/products/errors/product-error.enum.ts
export enum ProductErrorCode {
  PRODUCT_ALREADY_EXISTS = 'PRODUCT_ALREADY_EXISTS',
  PRODUCT_NOT_FOUND = 'PRODUCT_NOT_FOUND',
  PRODUCT_VARIANT_ALREADY_EXISTS = 'PRODUCT_VARIANT_ALREADY_EXISTS',
  PRODUCT_VARIANT_NOT_FOUND = 'PRODUCT_VARIANT_NOT_FOUND',
  PRODUCT_IMAGE_ALREADY_EXISTS = 'PRODUCT_IMAGE_ALREADY_EXISTS',
  PRODUCT_IMAGE_NOT_FOUND = 'PRODUCT_IMAGE_NOT_FOUND',
  PRODUCT_TAG_ALREADY_EXISTS = 'PRODUCT_TAG_ALREADY_EXISTS',
  PRODUCT_TAG_NOT_FOUND = 'PRODUCT_TAG_NOT_FOUND',
}